refactor: Move splash screen to WorkspaceProvider
