+++ source/app/service-providers/workspaces/splash-screen.ts
